Move ColorScale from MapConfig to MapData
Because it uses data to generate its output.
